ignored railtie and rails dir
